scripts/requirements.txt
python-dotenv>=1.0.0
ddgs>=1.0.0

